refactor: move datanode options to common_options
