+++ src/cmd/src/options.rs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_options::datanode::DatanodeOptions;
 use common_options::frontend::FrontendOptions;
 use common_options::meta::MetaSrvOptions;
 use common_telemetry::logging::LoggingOptions;
 use config::{Config, Environment, File, FileFormat};
-use datanode::datanode::DatanodeOptions;
 use serde::{Deserialize, Serialize};
 use snafu::ResultExt;
 
@@ -114,8 +114,8 @@ mod tests {
     use std::io::Write;
     use std::time::Duration;
 
+    use common_options::datanode::{DatanodeOptions, ObjectStoreConfig};
     use common_test_util::temp_dir::create_named_temp_file;
-    use datanode::datanode::{DatanodeOptions, ObjectStoreConfig};
 
     use super::*;
