+++ src/common/options/Cargo.toml
@@ -5,12 +5,17 @@ edition.workspace = true
 license.workspace = true
 
 [dependencies]
+common-base = { path = "../base" }
 common-error = { path = "../error" }
 common-grpc = { path = "../grpc" }
 common-procedure = { path = "../procedure" }
 common-telemetry = { path = "../telemetry" }
-servers = { path = "../../servers" }
+humantime-serde = "1.1"
+rustls = "0.21"
+rustls-pemfile = "1.0"
+secrecy = { version = "0.8", features = ["serde", "alloc"] }
 serde.workspace = true
 serde_json.workspace = true
+strum = { version = "0.24", features = ["derive"] }
 snafu.workspace = true
 toml.workspace = true
