+++ src/datanode/src/store/azblob.rs
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_options::datanode::AzblobConfig;
 use common_telemetry::logging::info;
 use object_store::services::Azblob as AzureBuilder;
 use object_store::{util, ObjectStore};
 use secrecy::ExposeSecret;
 use snafu::prelude::*;
 
-use crate::datanode::AzblobConfig;
 use crate::error::{self, Result};
 
 pub(crate) async fn new_azblob_object_store(azblob_config: &AzblobConfig) -> Result<ObjectStore> {
