+++ src/datanode/src/store/s3.rs
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_options::datanode::S3Config;
 use common_telemetry::logging::info;
 use object_store::services::S3 as S3Builder;
 use object_store::{util, ObjectStore};
 use secrecy::ExposeSecret;
 use snafu::prelude::*;
 
-use crate::datanode::S3Config;
 use crate::error::{self, Result};
 
 pub(crate) async fn new_s3_object_store(s3_config: &S3Config) -> Result<ObjectStore> {
