+++ src/servers/Cargo.toml
@@ -28,6 +28,7 @@ common-grpc-expr = { path = "../common/grpc-expr" }
 common-mem-prof = { path = "../common/mem-prof", optional = true }
 common-pprof = { path = "../common/pprof", optional = true }
 common-query = { path = "../common/query" }
+common-options = { path = "../common/options" }
 common-recordbatch = { path = "../common/recordbatch" }
 common-runtime = { path = "../common/runtime" }
 common-telemetry = { path = "../common/telemetry" }
