+++ src/storage/Cargo.toml
@@ -15,6 +15,7 @@ bytes = "1.1"
 common-base = { path = "../common/base" }
 common-datasource = { path = "../common/datasource" }
 common-error = { path = "../common/error" }
+common-options = { path = "../common/options" }
 common-query = { path = "../common/query" }
 common-recordbatch = { path = "../common/recordbatch" }
 common-runtime = { path = "../common/runtime" }
