+++ tests-integration/src/cluster.rs
@@ -23,11 +23,13 @@ use common_base::Plugins;
 use common_grpc::channel_manager::ChannelManager;
 use common_meta::peer::Peer;
 use common_meta::DatanodeId;
+use common_options::datanode::{DatanodeOptions, ObjectStoreConfig};
 use common_options::frontend::FrontendOptions;
 use common_options::meta::MetaSrvOptions;
+use common_options::servers::Mode;
 use common_runtime::Builder as RuntimeBuilder;
 use common_test_util::temp_dir::create_temp_dir;
-use datanode::datanode::{DatanodeOptions, ObjectStoreConfig, ProcedureConfig};
+use datanode::datanode::ProcedureConfig;
 use datanode::heartbeat::HeartbeatTask;
 use datanode::instance::Instance as DatanodeInstance;
 use frontend::instance::{FrontendInstance, Instance as FeInstance};
@@ -40,7 +42,6 @@ use meta_srv::service::store::kv::KvStoreRef;
 use meta_srv::service::store::memory::MemStore;
 use servers::grpc::GrpcServer;
 use servers::query_handler::grpc::ServerGrpcQueryHandlerAdaptor;
-use servers::Mode;
 use tonic::transport::Server;
 use tower::service_fn;
